requirements.txt
torch
torchvision
numpy
opencv-python
matplotlib
Pillow
scikit-image
